refactor(showcase): use plugin-specific page
