+++ src/plugins/apps/components/AppCard.module.scss
