+++ src/plugins/apps/components/apps.module.scss
